Refactor tests and command line arguments
